requirements.txt
pandas
numpy
matplotlib
opencv-python
torch
torchvision
Pillow
sklearn
joblib
scipy
